Django/requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.24
django-bootstrap3==11.1.0
django-braces==1.13.0
django-choices==1.7.0
django-cors-headers==3.1.0
django-decouple==2.1
django-filter==2.2.0
django-heroku==0.3.1
django-hitcount==1.3.1
django-leaflet==0.24.0
django-model-utils==3.2.0
django-multiupload==0.5.2
django-registration==2.4.1
django-rest-auth==0.9.5
django-star-ratings==0.8.0
django-tinymce==2.8.0
djangorestframework==3.10.2
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
pyuploadcare==2.6.0
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
swapper==1.1.1
urllib3==1.25.3
whitenoise==4.1.3
